refactor: move sessionIsActive to new useCases folder
